fix: separate `RegisterSchema` type and clear error message when age input is empty
