+++ app/(viewport)/form/page.tsx
@@ -6,7 +6,6 @@ import { useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import * as z from 'zod';
 
 import { postForm } from '@/api/postForm';
 import Step1 from '@/components/steps/Step1';
@@ -23,31 +22,7 @@ import {
   SelfPromotionCategories,
 } from '@/constants/categories';
 import { StepType, Steps } from '@/constants/steps';
-
-export const RegisterSchema = z.object({
-  mbti: z.string().min(1, { message: ErrorMessages.mbti }),
-  name: z.string().min(1, { message: ErrorMessages.name }),
-  gender: z.string().min(1, { message: ErrorMessages.gender }),
-  email: z.string().email(ErrorMessages.email),
-  instagramId: z.string(),
-  age: z.coerce
-    .number()
-    .min(20, { message: ErrorMessages.age })
-    .max(100)
-    .optional(),
-  type: z.string().optional(),
-  animal: z.string().min(1, { message: ErrorMessages.animal }),
-  situation: z
-    .array(
-      z.object({
-        scenario: z.number(),
-        choice: z.number(),
-      }),
-    )
-    .length(4),
-});
-
-export type RegisterSchemaType = z.infer<typeof RegisterSchema>;
+import { RegisterSchema, type RegisterSchemaType } from '@/lib/registerSchema';
 
 export default function Form() {
   const router = useRouter();
@@ -77,7 +52,10 @@ export default function Form() {
         return BasicInformationCategories;
       case 1:
         if (methods.getValues(Category.age)) return AdditionalCategories;
-        else return [];
+        else {
+          methods.clearErrors('age');
+          return [];
+        }
       case 2:
         return SelfPromotionCategories;
       case 3:
